Add new files for the other models and their respective layers of the api.

+++ src/main/java/nl/benzelinsky/techiteasybackend/repositories/TelevisionRepository.java
@@ -3,5 +3,5 @@ package nl.benzelinsky.techiteasybackend.repositories;
 import nl.benzelinsky.techiteasybackend.models.Television;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TelevisionRepository extends JpaRepository<Television, Integer> {
+public interface TelevisionRepository extends JpaRepository<Television, Long> {
 }
